+++ src/components/SidePane/HelpContent/settings/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import BulletList from '@components/SidePane/HelpBulletList';
-import ExpandableHelp from '@components/SidePane/HelpExpandable';
+import BulletList from '@components/SidePane/HelpComponents/HelpBulletList';
+import ExpandableHelp from '@components/SidePane/HelpComponents/HelpExpandable';
 import Text from '@components/Text';
 import type {ThemeStyles} from '@styles/index';
