+++ src/components/SidePane/HelpContent/settings/workspaces/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/no-unescaped-entities */
 import React from 'react';
-import BulletList from '@components/SidePane/HelpBulletList';
-import ExpandableHelp from '@components/SidePane/HelpExpandable';
+import BulletList from '@components/SidePane/HelpComponents/HelpBulletList';
+import ExpandableHelp from '@components/SidePane/HelpComponents/HelpExpandable';
 import Text from '@components/Text';
 import type {ThemeStyles} from '@styles/index';
