+++ src/components/SidePane/HelpModal/index.tsx
@@ -6,14 +6,14 @@ import {useOnyx} from 'react-native-onyx';
 // Modal from react-native can't be used here, as it would block interactions with the rest of the app
 import ModalPortal from 'react-native-web/dist/exports/Modal/ModalPortal';
 import FocusTrapForModal from '@components/FocusTrap/FocusTrapForModal';
-import SidePaneOverlay from '@components/SidePane/SidePaneOverlay';
+import HelpContent from '@components/SidePane/HelpComponents/HelpContent';
+import HelpOverlay from '@components/SidePane/HelpComponents/HelpOverlay';
 import useKeyboardShortcut from '@hooks/useKeyboardShortcut';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useSafeAreaPaddings from '@hooks/useSafeAreaPaddings';
 import useThemeStyles from '@hooks/useThemeStyles';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
-import HelpContent from './HelpContent';
 import type HelpProps from './types';
 
 function Help({sidePaneTranslateX, closeSidePane, shouldHideSidePaneBackdrop}: HelpProps) {
@@ -60,7 +60,7 @@ function Help({sidePaneTranslateX, closeSidePane, shouldHideSidePaneBackdrop}: H
                 <View style={styles.sidePaneContainer}>
                     <View>
                         {!shouldHideSidePaneBackdrop && (
-                            <SidePaneOverlay
+                            <HelpOverlay
                                 onBackdropPress={closeSidePane}
                                 isRHPVisible={isRHPVisible}
                             />
